[Fix] refine Responses API adapter handling
